feat: Add semaphore support for Linux

+++ semaphore.go
@@ -8,7 +8,7 @@ import (
 // tasks when its count reaches zero and allows any thread to increment
 // the count up to a pre-configured maximum value.
 type Semaphore interface {
-	Release(count int)
+	Release(count uint16)
 	Wait(timeout time.Duration) error
 	Close()
 }
